requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0 ; python_version >= '3.8'
anyio==3.7.1 ; python_version >= '3.7'
blinker==1.7.0 ; python_version >= '3.8'
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
distro==1.8.0 ; python_version >= '3.6'
exceptiongroup==1.1.3 ; python_version < '3.11'
flask==3.0.0
h11==0.14.0 ; python_version >= '3.7'
httpcore==1.0.1 ; python_version >= '3.8'
httpx==0.25.1 ; python_version >= '3.8'
idna==3.4 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
openai==1.1.1
pydantic==2.4.2 ; python_version >= '3.7'
pydantic-core==2.10.1 ; python_version >= '3.7'
python-dotenv==1.0.0
python-json-logger==2.0.7
requests==2.31.0
sniffio==1.3.0 ; python_version >= '3.7'
tqdm==4.66.1 ; python_version >= '3.7'
typing-extensions==4.8.0 ; python_version >= '3.8'
urllib3==2.0.7 ; python_version >= '3.7'
werkzeug==3.0.1 ; python_version >= '3.8'
